+++ travis/test.py
@@ -1,3 +1,9 @@
+import sys
+import os
+
 print("Hello World")
-assert 1 == 1
-assert 45 == 0
+print(sys.version[:6])
+if sys.version[:6] == "3.7":
+	assert 1
+else:
+	assert 0