+++ CHANGELOG.md
@@ -1,8 +1,12 @@
-## Indev 2.1b [Sep 27, 2019]
+#### Indev 2.1c [Sep 27, 2019]
+* Minor change to README
+* Testing Travis #3 (Last One)
+
+#### Indev 2.1b [Sep 27, 2019]
 * Minor change to README
 * Testing Travis #2
 
-## Indev 2.1a [Sep 27, 2019]
+#### Indev 2.1a [Sep 27, 2019]
 * Minor change to README
 * Testing Travis